inline memtablerep::KeyComparator
